+++ StandardLibrary/src/Index.dfy
@@ -5,7 +5,7 @@ include "../../libraries/src/MutableMap/MutableMap.dfy"
 include "./Actions.dfy"
 include "./Base64.dfy"
 include "./Base64Lemmas.dfy"
-include "./CallMany.dfy"
+include "./ConcurrentCall.dfy"
 include "./FloatCompare.dfy"
 include "./HexStrings.dfy"
 include "./Sets.dfy"
